requirements.txt
django==4.1.5
django-htmx==1.13.0
django-phonenumber-field[phonenumbers]==7.0.2
django-imagekit==4.1.0
Pillow==9.4.0